Insight = notes only
If the link to the gateway on the shortest path does not need to be cut immediately
(path > 2 nodes, note: path is [current node, next node]), 
then cut a gateway link that runs from the nearest node that has two gateway links (doesn't matter which link)
This adjustment should enable the solution to pass all tests (not implemented in this code commit). 

So we need a way of tracking how far away gateway node are , which we have since a path is created for each exit in skynet.getExits() and the actions string is currently set to the shortest path. 
But we really need to know which nodes have two gateway links. We could loop through all the exit paths and extract the 1 from last node, saving them in a list and finding those that appear more than once. We would still need to compute distance. Perhaps we could save them in a map along with their distance (path length -1).