Added a search for multiple gateway nodes (has serious flaws)
This implementation checks if the best path contains more than 2 nodes (in other words, it checks if is NOT necessary to sever a link immediately) and if so, it searches for nodes that have multiple gateway exit links. It finds the nearest of these nodes and severs one of the two exit links (it doesn't matter which one). 

This method fails, because the routine only searches for nodes that are on a BEST PATH to an exit node, not ANY path to an exit node. Consequently, it can miss nearby that nodes that have links that should be severed as a priority, in favour of more distant nodes.