Improved multi exit node search, but insufficient
By establishing, a list of multi exit nodes on initialisation, improved the multi exit node search to find the nearest multi exit node, even if it is not on a best path to an exit. This work quite well, but fails on complex meshes, where the nearest multi exit node is not always the best one to sever. The best one to sever is the one with the least "spare moves" on its critical path. In other words, sever it now or never. 
So, when finding the best path to each multi exit node, I could count the "spare moves" and select the one with the least. How to count "spare moves" ? See if each node in the path is the source of an exit node. If it isn't. there is a spare move.