Change multi-exit node selection to spareMove criteria - still insufficient
The selection of the multi-exit node to sever is now based on the lowest number of available 'spare moves', rather than the shortest possible path. 
However, this doesn't work for complex meshes (Test 5), because the spare move count is still being made on the shortest path to the multi-exit node, not the path that the agent will actually take (which perambulates as the agent heads towards exits that may be unconnected to the multi-exit nodes). 
It seems that I have to predict the path of the agent, from starting position to multi-exit node and count the spare moves on this path. This is hard and iterative. I have to calculate the path to the nearest exit, figure out when I would sever it and then calculate the next path to an exit, etc, until the agent reaches a multi-exit node. All the while counting the spare moves. Then repeat for any other multi-exit nodes. Need to think about whether there is a more efficient approach. 
Curiously, the solution does now work for Test 6.